refactor: prettify import paths
