+++ packages/digitraffic-mqtt/src/handlers/subscribe_to_train_locations.ts
@@ -1,14 +1,14 @@
-import type { MessageGeneratorResult } from '../base/message_generator.js'
-import type { GpsLocation } from '@junat/digitraffic/types/index.js'
+import type { MessageGeneratorResult } from '~/base/message_generator'
+import type { GpsLocation } from '@junat/digitraffic/types'
 
 import mqtt from 'mqtt'
 
-import { hasConnected } from '../base/has_connected.js'
-import { messageGenerator } from '../base/message_generator.js'
-import { createHandler, HandlerReturn } from '../base/create_handler.js'
-import { close } from '../base/close.js'
+import { hasConnected } from '~/base/has_connected'
+import { messageGenerator } from '~/base/message_generator'
+import { createHandler, HandlerReturn } from '~/base/create_handler'
+import { close } from '~/base/close'
 
-import { MQTT_URL } from '../constants/index.js'
+import { MQTT_URL } from '~/constants/index'
 
 export interface TrainLocationsMqttClient extends HandlerReturn {
   /**
