+++ packages/digitraffic-mqtt/tests/base/close.test.ts
@@ -1,6 +1,6 @@
 import { vi, it, expect } from 'vitest'
 
-import { close } from '../../src/base/close'
+import { close } from '~/base/close'
 
 it('calls client.end if client is not disconnected already', async () => {
   const client: any = {
