+++ packages/digitraffic-mqtt/tsconfig.json
@@ -13,7 +13,10 @@
     "skipLibCheck": true,
     "incremental": true,
     "strict": true,
-    "target": "es2020"
+    "target": "es2020",
+    "paths": {
+      "~/*": ["./src/*.js", "./src/*"]
+    }
   },
   "include": ["src"],
   "exclude": ["src/setup.ts", "src/mqtt_server_setup.ts"]
